arm: move gicv2 and gicv3 in separate folders
Preparing the hierarchy of the gic module
for adding snapshot support for gicv2.
