+++ src/arch/src/aarch64/gic/gicv3/mod.rs
@@ -1,15 +1,16 @@
 // Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+mod regs;
 use std::{boxed::Box, result};
 
+use crate::aarch64::gic::{Error, GICDevice};
 use kvm_ioctls::DeviceFd;
-
-use super::{Error, GICDevice};
+pub use regs::{restore_state, save_state, GicState};
 
 type Result<T> = result::Result<T, Error>;
 
-pub struct GICv3 {
+pub(crate) struct GICv3 {
     /// The file descriptor for the KVM device
     fd: DeviceFd,
 
@@ -113,5 +114,46 @@ impl GICDevice for GICv3 {
         )?;
 
         Ok(())
+    }
+}
+
+/// Function that flushes
+/// RDIST pending tables into guest RAM.
+///
+/// The tables get flushed to guest RAM whenever the VM gets stopped.
+fn save_pending_tables(fd: &DeviceFd) -> Result<()> {
+    let init_gic_attr = kvm_bindings::kvm_device_attr {
+        group: kvm_bindings::KVM_DEV_ARM_VGIC_GRP_CTRL,
+        attr: u64::from(kvm_bindings::KVM_DEV_ARM_VGIC_SAVE_PENDING_TABLES),
+        addr: 0,
+        flags: 0,
+    };
+    fd.set_device_attr(&init_gic_attr)
+        .map_err(|e| Error::DeviceAttribute(e, true, kvm_bindings::KVM_DEV_ARM_VGIC_GRP_CTRL))
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::aarch64::gic::create_gic;
+    use kvm_ioctls::Kvm;
+
+    #[test]
+    fn test_save_pending_tables() {
+        use std::os::unix::io::AsRawFd;
+
+        let kvm = Kvm::new().unwrap();
+        let vm = kvm.create_vm().unwrap();
+        let gic = create_gic(&vm, 1).expect("Cannot create gic");
+        assert!(save_pending_tables(&gic.device_fd()).is_ok());
+
+        unsafe { libc::close(gic.device_fd().as_raw_fd()) };
+
+        let res = save_pending_tables(&gic.device_fd());
+        assert!(res.is_err());
+        assert_eq!(
+            format!("{:?}", res.unwrap_err()),
+            "DeviceAttribute(Error(9), true, 4)"
+        );
     }
 }
